requirements.txt
numpy==1.24.4
pandas==2.0.3
pydantic==2.0.2
python-dotenv==0.21.1
requests==2.28.2
